weather_recognition/requirements.txt
torch>=2.0.0
torchvision>=0.15.0
img2vec-pytorch>=1.0.1
scikit-learn>=1.0.0
Pillow>=9.0.0
matplotlib>=3.5.0
joblib>=1.1.0
